Add duration support (#75)
* Understand hour and minute durations

Values in the format hh:mm are assumed to be durations expressed as
hours and minutes. They may also be used as a time of day - for example,
14:00 for 2pm).

Includes a new formatting directive ";hm" to specify output as hours and
minutes.

* Add timesheet example

Adds an example showing how one might use hour-and-minute duration
values as a time of day.